Translator: chaned way how sentence is capitalized
Previously sentence was capitalized based on full stops and first word
so only first word in a sentence was capitalized. Now since the word
order isn't changed in the translation capitalization isn't changed,
meaning that if a word is capitalized in source text it will also be
capitalized in translated text.